Transfer mocks into fakes utility
