+++ tests/tamr_client/test_project.py
@@ -1,13 +1,13 @@
 import pytest
 
 import tamr_client as tc
-from tests.tamr_client import fake, utils
+from tests.tamr_client import fake
 
 
 @fake.json
 def test_from_resource_id_mastering():
-    s = utils.session()
-    instance = utils.instance()
+    s = fake.session()
+    instance = fake.instance()
 
     project = tc.project.from_resource_id(s, instance, "1")
     assert isinstance(project, tc.MasteringProject)
@@ -17,8 +17,8 @@ def test_from_resource_id_mastering():
 
 @fake.json
 def test_from_resource_id_not_found():
-    s = utils.session()
-    instance = utils.instance()
+    s = fake.session()
+    instance = fake.instance()
 
     with pytest.raises(tc.project.NotFound):
         tc.project.from_resource_id(s, instance, "1")
